1、audio_stream_id 设计十分糟糕有bug，2、video 旋转问题

+++ app/src/main/cpp/FPlayer.cpp
@@ -89,7 +89,7 @@ int start_player(const char *src_filename, ANativeWindow *window,
     //
     ff_init();
     // ->>
-    ff_sec_duration = fmt_ctx->duration / AV_TIME_BASE; //ms
+    ff_sec_duration = (int32_t)fmt_ctx->duration / AV_TIME_BASE; //ms
     if (ff_sec_duration <= 0) {
         ff_sec_duration = 1; //ms
     }
